+++ examples/univariate_return_simulation_example.py
@@ -1,29 +1,10 @@
-from datetime import datetime
 import math
-import typing as tp
 
-import arch
 import matplotlib.pyplot as plt
-import numpy as np
 import os
-import pandas as pd
-import pyarrow
 from recombinator import stationary_bootstrap
-import scipy as sp
-import statsmodels.api as sm
-from statsmodels.graphics.gofplots import qqplot_2samples
-from statsmodels.tsa.stattools import adfuller
 
-from ovm.garch_estimation import estimate_garch_parameters
-
-from ovm.historical_data_io import (
-    PriceHistoryColumnNames as PHCN,
-    compute_scaling_factor,
-    compute_scaled_log_returns,
-    PriceHistory,
-    compute_number_of_days_in_price_history,
-    compute_log_returns_from_price_history,
-    save_price_histories,
+from ovm.historical.data_io import (
     load_price_history
 )
